refactor(kubernetes): Push namespace logic down to V1 and V2
A following commit will simplify the logic for handling namespaces
in the kubernetes caching agents. There is a lot of legacy V1 code
in groovy that depends on the particular implementation that I don't
want to refactor and test as part of this change. In order to restrict
the change to V2, push this logic down from the base caching agent to
the V1 and V2 caching agents, where it is for now duplicated but will
soon be changed in V2.

This also makes the KubernetesMetricCachingAgent extend the
KubernetesV2CachingAgent rather than the base KubernetesCachingAgent
so it can share the new namespace functionality.
